refactoring spec tests
